# docs/swagger.yaml
definitions:
  app.Pager:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_rows:
        type: integer
    type: object
  errcode.Error:
    type: object
  model.Article:
    properties:
      content:
        type: string
      cover_image_url:
        type: string
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      desc:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      state:
        type: integer
      title:
        type: string
    type: object
  model.ArticleSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Article'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
    type: object
  model.Tag:
    properties:
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      name:
        type: string
      state:
        type: integer
    type: object
  model.TagSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
    type: object
info:
  contact: {}
  description: A simple blog service
  termsOfService: https://github.com/lachimere
  title: Blog Service
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: Article Name
        in: query
        name: name
        type: string
      - description: Tag ID
        in: query
        name: tag_id
        type: integer
      - description: State
        in: query
        name: state
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.ArticleSwagger'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: List articles
    post:
      parameters:
      - description: Tag ID
        in: body
        name: tag_id
        required: true
        schema:
          type: string
      - description: Article Title
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: Article Description
        in: body
        name: desc
        schema:
          type: string
      - description: Cover Image URL
        in: body
        name: cover_image_url
        required: true
        schema:
          type: string
      - description: Article Content
        in: body
        name: content
        required: true
        schema:
          type: string
      - description: Creator
        in: body
        name: created_by
        required: true
        schema:
          type: integer
      - description: State
        in: body
        name: state
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Create article
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Delete article
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Get an article
    put:
      parameters:
      - description: Tag ID
        in: body
        name: tag_id
        schema:
          type: string
      - description: Article Title
        in: body
        name: title
        schema:
          type: string
      - description: Article Description
        in: body
        name: desc
        schema:
          type: string
      - description: Cover Image URL
        in: body
        name: cover_image_url
        schema:
          type: string
      - description: Article Content
        in: body
        name: content
        schema:
          type: string
      - description: Modifier
        in: body
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Update article
  /api/v1/tags:
    get:
      parameters:
      - description: Tag Name
        in: query
        maxLength: 100
        name: name
        type: string
      - default: 1
        description: State
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.TagSwagger'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: List tags
    post:
      parameters:
      - description: Tag Name
        in: body
        maxLength: 100
        minLength: 3
        name: name
        required: true
        schema:
          type: string
      - default: 1
        description: State
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: Creator
        in: body
        maxLength: 100
        minLength: 3
        name: created_by
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Create tag
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Delete tag
    put:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag Name
        in: body
        maxLength: 100
        minLength: 3
        name: name
        schema:
          type: string
      - default: 1
        description: State
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: Modifier
        in: body
        maxLength: 100
        minLength: 3
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/model.Tag'
            type: array
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: Update tag
swagger: "2.0"